Home

PythonToJuliaHelper.jl

PythonToJuliaHelper.PYTHON2JULIA_WORDS

PythonToJuliaHelper.add_end_to_block!

PythonToJuliaHelper.anonymous_function

PythonToJuliaHelper.delete_colon!

PythonToJuliaHelper.function_docstrings

PythonToJuliaHelper.get_indentation

PythonToJuliaHelper.move_all_docstrings

PythonToJuliaHelper.python2julia

PythonToJuliaHelper.ranges

PythonToJuliaHelper.PYTHON2JULIA_WORDS — Constant.

List of Pairs to replace Python's words by Julia's words. In the Regex,  indicates a word boundary.

source

PythonToJuliaHelper.python2julia — Method.

It takes Python code lines and transform it to something closer to Julia code.

It moves docstrings before their function definitions.

It adds an end line to for, if and function blocks.

It deletes colons at the end of for, if, def, else and elif statements.

It changes anonymous function syntax.

It changes (x)range syntax without modification of the values.

It replaces python words using the pairs in PYTHON2JULIA_WORDS.

source

PythonToJuliaHelper.add_end_to_block! — Method.

It adds an end line at the end of an indented block.

source

PythonToJuliaHelper.anonymous_function — Method.

Python | lambda x, y : y + x * 2 Julia | (x,y) -> y + x * 2

julia> using PythonToJuliaHelper

julia> PythonToJuliaHelper.anonymous_function("lambda x, y : y + x * 2")
"(x, y) -> y + x * 2"

julia> PythonToJuliaHelper.anonymous_function("lambda x : x * 2")
"x -> x * 2"

source

PythonToJuliaHelper.delete_colon! — Method.

It deletes the colon at the end of the line.

source

PythonToJuliaHelper.function_docstrings — Method.

It moves the docstring before the function definition.

source

PythonToJuliaHelper.get_indentation — Method.

Return the indentation of the line.

julia> using PythonToJuliaHelper

julia> PythonToJuliaHelper.get_indentation("        if value:")
"        "

julia> PythonToJuliaHelper.get_indentation("def fun(arg):")
""

source

PythonToJuliaHelper.move_all_docstrings — Method.

It moves all the docstrings before their function definitions.

source

PythonToJuliaHelper.ranges — Method.

Python | range(stop) Python | range(start, stop) Python | range(start, stop, step) Julia | start:stop Julia | start:step:stop

Python's range(3) and range(0, 3) (i.e. 0,1,2) are going to be Julia's 0:3 (i.e. 0,1,2,3)

Python's range and xrange are translated to ranges unless collect=true.

julia> using PythonToJuliaHelper

julia> PythonToJuliaHelper.ranges("range(stop)")
"0:stop"

julia> PythonToJuliaHelper.ranges("range(start, stop)")
"start:stop"

julia> PythonToJuliaHelper.ranges("range(start,stop,step)")
"start:step:stop"

julia> PythonToJuliaHelper.ranges("range(start,stop,step)", collect=true)
"collect(start:step:stop)"

julia> PythonToJuliaHelper.ranges("xrange(start,stop,step)", collect=true)
"start:step:stop"

source